extend PO037 to perform more hygiene checks
