+++ test/fixtures/resources/PO037/datacapture1/apiproxy/policies/DC-Collect-Source-with-attrs.xml
@@ -1,0 +1,15 @@
+<DataCapture name="DC-Collect-Source-with-attrs">
+  <IgnoreUnresolvedVariables>false</IgnoreUnresolvedVariables>
+  <ThrowExceptionOnLimit>false</ThrowExceptionOnLimit>
+
+  <Capture>
+    <DataCollector>dc_req_headers_length</DataCollector>
+    <Collect ref="computed-auth-header" default='none'>
+      <!-- Source does not support attributes -->
+      <Source attr1='goo'>request</Source>
+      <Header name="my_auth">
+        <Pattern ignoreCase="false">Bearer {$}</Pattern>
+      </Header>
+    </Collect>
+  </Capture>
+</DataCapture>
