+++ test/fixtures/resources/PO037/datacapture1/apiproxy/policies/DataCapture-QueryParam-with-message-Source-1.xml
@@ -1,0 +1,22 @@
+<DataCapture name="DataCapture-QueryParam-with-message-Source-1">
+  <IgnoreUnresolvedVariables>false</IgnoreUnresolvedVariables>
+  <ThrowExceptionOnLimit>false</ThrowExceptionOnLimit>
+
+  <!-- Existing Variable -->
+  <Capture>
+    <DataCollector>dc_req_headers_length</DataCollector>
+    <Collect ref="computed-headers-length" default="0"/>
+  </Capture>
+
+  <!-- QueryParam -->
+  <Capture>
+    <DataCollector>dc_path_segment</DataCollector>
+    <Collect default="none">
+      <!-- explicitly specify Source -->
+      <Source>message</Source>
+      <QueryParam name='code'>
+        <Pattern ignoreCase="false">{$}</Pattern>
+      </QueryParam>
+    </Collect>
+  </Capture>
+</DataCapture>
